requirements.txt
APScheduler==3.6.3
astral==1.10.1
Click==7.0
Flask==1.1.1
Flask-API==2.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
tzlocal==2.0.0
Werkzeug==0.16.1
